:truck: Move effects to their own dirs
