+++ demo/index.ts
@@ -12,10 +12,8 @@ import {
   Screenshot,
   Crop,
   Rotate,
-  Filters,
   Flip,
   Blur,
-  Sharpen,
   Flickr,
   NineGag,
   Pinterest,
